+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 
 - Changed Maven coordinates from `dev.orne.test:generators` to `dev.orne.test:orne-test-generators`.
 
+### Added
+
+- Expose "generator not found" generator placeholder `Generator.MISSING`.
+
 ## dev.orne.test:generators - 0.2.0 - 2023-12-14
 
 ### Added